# Use a faster implementation of AliasTables