--- a/src/samplers/aliastable.jl
+++ b/src/samplers/aliastable.jl
@@ -1,4 +1,3 @@
-using AliasTables
 struct AliasTable <: Sampleable{Univariate,Discrete}
     at::AliasTables.AliasTable{UInt64, Int}
     AliasTable(probs::AbstractVector{<:Real}) = new(AliasTables.AliasTable(probs))