--- a/test/univariate/discrete/categorical.jl
+++ b/test/univariate/discrete/categorical.jl
@@ -93,9 +93,9 @@ end
 end
 
 @testset "reproducibility across julia versions" begin
-    d= Categorical([0.1, 0.2, 0.7])
+    d = Categorical([0.1, 0.2, 0.7])
     rng = StableRNGs.StableRNG(600)
-    @test rand(rng, d, 10) == [2, 1, 3, 3, 2, 3, 3, 3, 3, 3]
+    @test rand(rng, d, 10) == [3, 1, 1, 2, 3, 2, 3, 3, 2, 3]
 end
 
 @testset "comparisons" begin