@devmotion 2024-04-08
Given the performance improvements, shouldn't this be used in or even contributed to StatsBase?